// frontend - src/app/model/takmicenje.ts
import { Takmicar } from "./takmicar";

export class Takmicenje{
    sport: string;
    disciplina: string;
    pol: string;
    datumPocetka: string;
    datumKraja: string;
    lokacija: string;
    tip: string;
    takmicari: Array<Takmicar>;
    delegat: string;
    status: string;
}